Internal-Knowledge-Base-Assistance-new/requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
openai==1.46.0
pinecone-client==3.2.2
jinja2==3.1.4
starlette==0.38.5
pydantic==2.9.2
httpx==0.27.2
